fix(electrical): rank severity so alarm outranks alert (#351)
The shared resolveMostSevereState short-circuited Alert > Alarm > Warn > Normal,
inverting the Signal K severity order (normal < nominal < alert < warn < alarm <
emergency). A device with a co-occurring alert and alarm rendered the lower-
severity alert colour, visually downgrading the alarm. Replace the short-circuit
with the canonical rank fold (returns the highest-ranked state, handling nominal
and emergency the old body dropped), fold solar-charger's identical private copy
onto the shared helper, and flip the spec assertion that codified the inversion.
